# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Book My Cleaning — AI receptionist for Jobber cleaning companies
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: The company's client directory, A→Z (owner/dispatcher only)
      responses:
        "200":
          description: Every known client for this company
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientRecord"
    post:
      operationId: createClient
      tags: [clients]
      summary: Add a client contact (owner/dispatcher only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRecord"
  /clients/{id}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client contact (owner/dispatcher only)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientRecord"
        "404": { description: Not found }
...
components:
  schemas:
...
    ClientRecord:
      type: object
      required: [id, name, source, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        # As typed/received; may be blank for a Jobber client with no number.
        phone: { type: ["string", "null"] }
        # Normalized +E.164 when the number was dialable, for digit search.
        phoneE164: { type: ["string", "null"] }
        tag:
          type: ["string", "null"]
          enum: [client, good_lead, bad_lead, spam, null]
        email: { type: ["string", "null"] }
        streetAddress: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        province: { type: ["string", "null"] }
        postalCode: { type: ["string", "null"] }
        # Set once the customer exists in Jobber too.
        jobberClientId: { type: ["string", "null"] }
        # Where this customer was first seen: booking | jobber | lead.
        source: { type: string }
        createdAt: { type: string, format: date-time }
    ClientInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        streetAddress: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        province: { type: ["string", "null"] }
        postalCode: { type: ["string", "null"] }
    ClientUpdate:
      type: object
      minProperties: 1
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        streetAddress: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        province: { type: ["string", "null"] }
        postalCode: { type: ["string", "null"] }
...